migrating existing tests into the functional area
