--- test_infra/fixtures/.gitignore
@@ -1,1 +1,0 @@
-*.json
